+++ nessi/io/su_fmt.py
@@ -9,27 +9,27 @@
 # ------------------------------------------------------------------
 """
 Support of Seismic Unix format and some commands.
-
-:copyright:
-    Damien Pageot (nessi.develop@protonmail.com)
-:license:
-    GNU Lesser General Public License, Version 3
-    (https://www.gnu.org/copyleft/lesser.html)
 """
 
+# Import __future__ for python2/3 compatibility
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import os, sys
-import numpy as np
-import matplotlib.pyplot as plt
+# Import python modules
+import os
+import sys
 import copy
+import numpy as np
+from scipy.signal import resample
+import matplotlib.pyplot as plt
 
+# Import fonctions from NeSSI package
 from nessi.signal import time_window
 from nessi.signal import space_window
 from nessi.signal import taper1d
 from nessi.signal import sin2filter
+
 
 class SUdata():
     """
@@ -37,10 +37,8 @@ class SUdata():
     """
 
     def __init__(self):
-        """Define the Seismic Unix header.
-
-        Args:
-            None
+        """
+        Define the Seismic Unix header.
         """
         self.sutype = np.dtype([
             ('tracl', np.int32), ('tracr', np.int32), \
@@ -78,39 +76,20 @@ class SUdata():
             ('timebas', np.int16), ('trwf', np.int16), \
             ('grnors', np.int16), ('grnofr', np.int16), \
             ('grnlof', np.int16), ('gaps', np.int16), \
-            ('otrav', np.int16), ('cdpx', np.int32), \
-            ('cdpy', np.int32), ('Inline3D', np.int32), \
-            ('Crossline3D', np.int32), ('ShotPoint', np.int32), \
-            ('ShotPointScalar', np.int16), \
-            ('TraceValueMeasurementUnit', np.int16), \
-            ('TransductionConstantMantissa', np.int32), \
-            ('TransductionConstantPower', np.int16), \
-            ('TransductionUnit', np.int16), \
-            ('TraceIdentifier', np.int16), ('ScalarTraceHeader', np.int16), \
-            ('SourceType', np.int16), \
-            ('SourceEnergyDirectionMantissa', np.int32), \
-            ('SourceEnergyDirectionExponent', np.int16), \
-            ('SourceMeasurementMantissa', np.int32), \
-            ('SourceMeasurementExponent', np.int16), \
-            ('SourceMeasurementUnit', np.int16), \
-            ('UnassignedInt1', np.int32), \
-            ('UnassignedInt2', np.int32)])
-
+            ('otrav', np.int16), ('d1', np.float32),\
+            ('f1', np.float32), ('d2', np.float32), \
+            ('f2', np.float32), ('ungpow', np.float32), \
+            ('unscale', np.float32), ('ntr', np.int32), \
+            ('mark', np.int16), ('shortpad', np.int16), \
+            ('unassignedInt1', np.int32), ('unassignedInt2', np.int32), \
+            ('unassignedInt3', np.int32), ('unassignedInt4', np.int32), \
+            ('unassignedFloat1', np.float32), ('unassignedFloat2', np.float32), \
+            ('unassignedFloat3', np.float32)])
 
         self.filename = ' '
         self.header = []
         self.trace = []
         self.endian = 'l'
-
-        # For FFT and MASW
-        self.nw = 0.
-        self.dw = 0.
-        self.fmin = 0.
-
-        # For MASW only
-        self.nv = 0.
-        self.dv = 0.
-        self.vmin = 0.
 
     def _check_endian(self):
         """
@@ -131,63 +110,98 @@ class SUdata():
 
 
     def read(self, filename, endian=' '):
-        """Read Seismic Unix file.
-
-        :param filename: name of the SU file
-        :param endian: byte order: little endian (default) 'l', big endian 'b'.
         """
+        Read Seismic Unix file.
+
+        :param filename: name of the SU file to read
+        :param endian: byte order: little endian 'l', big endian 'b'.
+        """
+
+        # This method can be easily simplified... for [0.3.0] version ?
         self.filename = filename
+
         if endian == ' ':
+            # Automatic checking of endianess
             self._check_endian()
 
+        # Open the file to read
         file = open(filename, 'rb')
+
+        # Get the header of the first trace (240 bytes)
         bhdr = file.read(240)
+
         if self.endian == 'b':
+            # Read the header of the first trace
             hdr = np.frombuffer(bhdr, dtype=self.sutype.newbyteorder(), count=1)[0]
+            # Get the  the first trace data values
             btrc = file.read(hdr['ns']*4)
             trc = np.frombuffer(btrc, dtype=('>f4', hdr['ns']), count=1)[0]
+            # Save the header and the trace
             self.header.append(hdr)
             self.trace.append(trc)
+            # Loop over traces until end of file
             EOF = False
             while EOF == False:
                 try:
+                    # Get header and trace
                     bhdr = file.read(240)
                     btrc = file.read(hdr['ns']*4)
                     hdr = np.frombuffer(bhdr, dtype=self.sutype.newbyteorder(), count=1)[0]
                     trc = np.frombuffer(btrc, dtype=('>f4', hdr['ns']), count=1)[0]
+                    # Save
                     self.header.append(hdr)
                     self.trace.append(trc)
                 except:
                     EOF = True
+            # Convert in numpy array format
             self.trace = np.array(self.trace)
             self.header = np.array(self.header)
 
         if self.endian == 'l':
+            # Read the header of the first trace
             hdr = np.frombuffer(bhdr, dtype=self.sutype, count=1)[0]
+            # Get the  the first trace data values
             btrc = file.read(hdr['ns']*4)
             trc = np.frombuffer(btrc, dtype=('<f4', hdr['ns']), count=1)[0]
+            # Save the header and the trace
             self.header.append(hdr)
             self.trace.append(trc)
+            # Loop over traces until end of file
             EOF = False
             while EOF == False:
                 try:
+                    # Get header and trace
                     bhdr = file.read(240)
                     btrc = file.read(hdr['ns']*4)
                     hdr = np.frombuffer(bhdr, dtype=self.sutype, count=1)[0]
                     trc = np.frombuffer(btrc, dtype=('<f4', hdr['ns']), count=1)[0]
+                    # Save the header and the trace
                     self.header.append(hdr)
                     self.trace.append(trc)
                 except:
                     EOF = True
+            # Convert in numpy array format
             self.header = np.array(self.header)
             self.trace = np.array(self.trace)
 
 
-    def image(self, bclip=None, wclip=None, clip=None, legend=0, label1=' ',
-              label2=' ', title=' '):
+    def image(self, key='tracl', bclip=None, wclip=None, clip=None, legend=0, label1=' ',
+              label2=' ', title=' ', cmap='gray'):
         """
-        matplotlib.pyplot.imshow adapted for SU files
+        matplotlib.pyplot.imshow adapted to plot SU files
+
+        :param key: header keyword (default tracl)
+        :param bclip: data values outside of [bclip,wclip] are clipped
+        :param wclip: data values outside of [bclip,wclip] are clipped
+        :param clip: clip used to determine bclip and wclip
+        :param legend: colorbar 0=no colorbar (default) 1=colorbar
+        :param label1: x-axis label
+        :param label2: y-axis label
+        :param title: title of the image
+        :param cmap: color map (defautl 'gray'): gray, jet, ...
         """
+
+        # Check clip, bclip and wclip
         if(clip == None and bclip == None and clip == None):
             bclip = np.amin(self.trace)
             wclip = np.amax(self.trace)
@@ -196,19 +210,113 @@ class SUdata():
                 bclip = -1.*clip
                 wclip = clip
 
-        t0 = float(self.header[0]['delrt'])/1000.
-        t1 = float(self.header[0]['ns']-1)*float(self.header[0]['dt'])/1000000.+t0
+        # Get ns and dt from header
+        ns = self.header[0]['ns']
+        dt = float(self.header[0]['dt']/1000000.)
+        if dt != 0:
+            y0 = float(self.header[0]['delrt'])/1000.
+            y1 = float(ns-1)*dt+y0
+            x0 = self.header[0][key]
+            x1 = self.header[-1][key]
 
-        plt.xlabel(label2)
-        plt.ylabel(label1)
+        if self.header[0]['trid'] == 118:
+            # Get d1
+            d1 = float(self.header[0]['d1'])
+            y0 = 0.
+            y1 = float(ns-1)*d1
+            x0 = self.header[0][key]
+            x1 = self.header[-1][key]
+
+        if self.header[0]['trid'] == 122:
+            # Get d1
+            d1 = float(self.header[0]['d1'])
+            y0 = 0.
+            y1 = float(ns-1)*d1
+            # Get d2
+            d2 = float(self.header[0]['d2'])
+            x0 = float(self.header[0]['f2'])
+            x1 = x0+float(len(self.header)-1)*d2
+
+        # Add labels to axes
+        plt.xlabel(label1)
+        plt.ylabel(label2)
+
+        # Add title to axis
         plt.title(title)
-        plt.imshow(self.trace.swapaxes(1,0), aspect='auto', cmap='gray',
-                   extent=[0., len(self.trace), t1, t0],
+
+        # Plot surface
+        plt.imshow(self.trace.swapaxes(1,0), aspect='auto', cmap=cmap,
+                   extent=[x0, x1, y1, y0],
                    vmin=bclip, vmax=wclip)
+
+        # Add legend
         if legend == 1:
             plt.colorbar()
 
-    def kill(self, key=' ', a=1, min=-1, count=1):
+    def wiggle(self, clip=-1., key='tracl', label1=' ', label2=' ', title=' ', tracecolor='black', tracestyle='-', skip=1, xcur=1):
+        """
+        Wiggle for SU files
+
+        :param clip: clip used to determine outside values to be clipped [-clip, clip]
+        :param key: header keyword (default tracl)
+        :param label1: x-axis label
+        :param label2: y-axis label
+        :param title: title of the image
+        :param tracecolor: color of the traces
+        :param tracestyle: style of the traces ('--', ':', ...)
+        :param skip: number of traces to skip for each plotted trace
+        :param xcur: factor to increase trace amplitudes on output
+        """
+
+        # Get ns and dt from header
+        ns = self.header[0]['ns']
+        dt = float(self.header[0]['dt']/1000000.)
+        ntrac = len(self.header)
+        if dt != 0:
+            y0 = float(self.header[0]['delrt'])/1000.
+            y1 = float(ns-1)*dt+y0
+            x0 = self.header[0][key]
+            x1 = self.header[-1][key]
+            d2 = 1.
+
+        if self.header[0]['trid'] == 118:
+            # Get d1
+            d1 = float(self.header[0]['d1'])
+            y0 = 0.
+            y1 = float(ns-1)*d1
+            x0 = self.header[0][key]
+            x1 = self.header[-1][key]
+
+        if self.header[0]['trid'] == 122:
+            # Get d1
+            d1 = float(self.header[0]['d1'])
+            y0 = 0.
+            y1 = float(ns-1)*d1
+            # Get d2
+            d2 = float(self.header[0]['d2'])
+            x0 = float(self.header[0]['f2'])
+            x1 = x0+float(len(self.header)-1)*d2
+
+        # Add labels
+        plt.xlabel(label1)
+        plt.ylabel(label2)
+
+        # Add axes
+        plt.title(title)
+
+        # Get the normalization parameter (for output)
+        y = np.linspace(y0, y1, ns)
+        if clip >= 0. :
+            norm = clip
+        else:
+            norm = np.amax(np.abs(self.trace))
+
+        # Plot the traces
+        for itrac in range(0, ntrac, skip):
+            wig = self.trace[itrac]/norm*d2*float(skip-1)*xcur
+            plt.plot(wig+x0+float(itrac)*d2, y, color=tracecolor, linestyle=tracestyle)
+
+    def kill(self, key=' ', a=1, min=0, count=1):
         """
         Zero out traces.
         If min= is set it overrides selecting traces by header.
@@ -223,13 +331,14 @@ class SUdata():
         dobskill = copy.deepcopy(self)
 
         # Get the number of traces
-        ntrac = self.traces.shape[0]
+        ntrac = self.trace.shape[0]
 
-        # Kill traces
-        if min > 0:
+        # Kill traces from min to min+icount
+        if key == ' ':
             for icount in range(0, count):
                 if min+icount < ntrac:
                     dobskill.trace[min+icount, :] = 0.
+        # Kill traces with the given header value
         else:
             if key != ' ':
                 for itrac in range(0, ntrac):
@@ -248,6 +357,7 @@ class SUdata():
         :param tmin: minimum time to pass (=0)
         :param tmax: maximum time to pass (=0)
         """
+
         # Create a copy of the input SU data
         dobsw = copy.deepcopy(self)
 
@@ -352,18 +462,35 @@ class SUdata():
         """
         Write SU file on disk
         """
+
+        # Open binary file
         file = open(filename, 'wb')
+
+        # Loop over traces
         for ir in range(0, len(self.header)):
+            # Write header
             file.write(self.header[ir])
+            # Write data
             file.write(self.trace[ir,:])
+
+        # Close the binary file
         file.close()
 
     def masw(self, vmin=0., vmax=1000., dv=5., fmin=1., fmax=100.):
         """
-        Calculate the dispersion diagram using MASW method
+        Calculate the dispersion diagram using MASW method.
+        Return the dispersion diagram, the velocity vector and the frequency vector.
+
+        :param vmin: minimum value to consider for the dispersion diagram
+        :param vmax: maximum value to consider for the dispersion diagram
+        :param dv: velocity sampling
+        :param fmin: minimum frequency to consider
+        :param fmax: maximum frequency to consider
         """
 
-        # Get offset
+        # Conversion to cython for performance (?)
+
+        # Get potential scaling factor on coordinates from header
         scalco = self.header[0]['scalco']
         if scalco < 0:
             scale = -1./float(scalco)
@@ -372,32 +499,223 @@ class SUdata():
         if scalco > 0:
             scale = float(scalco)
 
+        # Calculate offsets
         x = self.header[:]['sx']*scale-self.header[:]['gx']*scale
         y = self.header[:]['sy']*scale-self.header[:]['gy']*scale
         offset = np.sqrt(x**2+y**2)
-        # Velocity vector
+
+        # Create the velocity vector
         nv = int((vmax-vmin)/dv)+1
         vel = np.linspace(vmin, vmax, nv, dtype=np.float32)
 
-        # FFT
+        # Get the number of samples and the time sampling from header
         ns = int(self.header[0]['ns'])
         dt = self.header[0]['dt']/1000000.
+
+        # Apply Real Fourier transform to data
         gobs = np.fft.rfft(self.trace, axis=1)
+
+        # Get the corresponding frequency vector
         freq = np.fft.rfftfreq(ns, d=dt)
         dw = freq[1]
         iwmin = int(fmin/dw)
         nw = int((fmax-fmin)/dw)+1
 
-        #disp = cy.cmasw(gobs, iwmin, nw, offset, vel, freq)
-        # MASW
+        # Initialize temporary and dispersion diagram arrays
         tmp = np.zeros(nw, dtype=np.complex64)
         disp = np.zeros((nv, nw), dtype=np.float32)
+
+        # Loop over velocities
         for iv in range(0, nv):
             tmp[:] = complex(0., 0.)
+            # Loop over traces
             for ir in range(0, len(offset)):
+                # Loop over frequencies
                 for iw in range(0, nw):
+                    # Calculate the phase
                     phase = complex(0., 1.)*2.*np.pi*offset[ir]*freq[iw+iwmin]/vel[iv]
+                    # Stack over frequencies and receivers
                     tmp[iw] += gobs[ir, iw+iwmin]*np.exp(phase)
+            # Stack over velocities
             disp[iv,:] += np.abs(tmp[:])
 
         return disp, vel, freq[iwmin:iwmin+nw]
+
+    def resamp(self, nso, dto):
+        """
+        Resample data in time.
+        Based on scipy.signal.resample
+
+        :param nso: number of time samples in output
+        :param dto: time sampling in output
+        """
+
+        # Create a copy of the input SU data
+        dobsresamp = copy.deepcopy(self)
+
+        # Get values from header
+        ns = self.header[0]['ns']
+        dt = self.header[0]['dt']/1000000.
+
+        # Calculate time lenght for the old data
+        t_old = float(ns-1)*dt
+
+        # Calculate time lenght for the resampled data
+        t_resamp = float(nso-1)*dto
+
+        # Calculate the number of time samples of the old trace to resample
+        nsamp = int(t_resamp/dt)+1
+
+        # Resampling
+        if nsamp > ns:
+            print('Impossible to resample \n')
+        else:
+            if np.ndim(self.trace) == 1:
+                dobsresamp.trace = resample(self.trace[:,:nsamp], num=nso)
+            else:
+                dobsresamp.trace = resample(self.trace[:,:nsamp], num=nso, axis=1 )
+
+        # Edit header
+        dobsresamp.header[:]['ns'] = nso
+        dobsresamp.header[:]['dt'] = int(dto*1000000.)
+
+        return dobsresamp
+
+    def mute(self, xmute, tmute, key='tracl', ntaper=0, mode=0):
+        """
+        Mute above or below a user-defined polygonal lineself.
+
+        :param key: SU header key
+        :param xmute: array of position values
+        :param tmute: array of time values
+        :param ntaper: number of points to taper before mute
+        :param mode: mute below (0) or above (1)
+        """
+
+        # Create a copy of the input SU data
+        dobsmute = copy.deepcopy(self)
+
+        # Get values from header
+        ns = self.header[0]['ns']
+        dt = self.header[0]['dt']/1000000.
+        delrt = self.header[0]['delrt']/1000.
+        ntrac = len(self.header)
+
+        # Get the number of points
+        npts = len(xmute)
+
+        # Get the keyword values
+        keyval = self.header[:][key]
+
+        # Get trace number for each point
+        keytracn = np.zeros(npts, dtype=np.int)
+        for ipts in range(0, npts):
+            keytracn[ipts] = np.argmin(np.abs(xmute[ipts]-keyval[:]))
+
+        # Build the sine squared taper
+        taper = np.zeros(ntaper, dtype=np.float32)
+        for itaper in range(0, ntaper):
+            ftap = np.sin(float(itaper+1)*np.pi/(float(2*ntaper)))
+            taper[itaper] = ftap**2
+
+        # Build the polygonal line
+        polyline = np.zeros(ntrac, dtype=np.int)
+        ## First point
+        if keytracn[0] > 0:
+            polyline[:keytracn[0]] = int((tmute[0]-delrt)/dt)
+        ## Last point
+        if keytracn[-1] < ntrac:
+            polyline[keytracn[-1]:] = int((tmute[-1]-delrt)/dt)
+        ## Middle points
+        for ipts in range(1, npts):
+            slope = (tmute[ipts]-tmute[ipts-1])/(xmute[ipts]-xmute[ipts-1])
+            origin = tmute[ipts-1]
+            i = 0
+            for itrac in range(keytracn[ipts-1], keytracn[ipts]):
+                polyline[itrac] = int((slope*float(i)+tmute[ipts-1]-delrt)/dt)
+                i += 1
+
+        # Mute
+        if mode == 0: # Mute above
+            for itrac in range(0, ntrac):
+                imute = polyline[itrac]
+                # Apply taper
+                for j in range(0, ntaper):
+                    if imute-j >= 0:
+                        dobsmute.trace[itrac, imute-j] *= taper[ntaper-j-1]
+                if imute-ntaper >= 0:
+                    dobsmute.trace[itrac,:imute-ntaper] = 0.
+        if mode == 1: # Mute below
+            for itrac in range(0, ntrac):
+                imute = polyline[itrac]
+                # Apply taper
+                for j in range(0, ntaper):
+                    if imute+j < ns:
+                        dobsmute.trace[itrac, imute+j] *= taper[j]
+                if imute+ntaper < ns:
+                    dobsmute.trace[itrac,imute+ntaper:] = 0.
+
+        return dobsmute
+
+    def specfx(self):
+        """
+        Fourier spectrum (time to frequency) of traces using the numpy.fft functions.
+        """
+
+        # Create a copy of the input SU data
+        dobsspecfx = copy.deepcopy(self)
+
+        # Amplitude of the real Fourier transform
+        dobsspecfx.trace = np.absolute(np.fft.rfft(self.trace, axis=1))
+
+        # Get the frequency vector
+        ns = dobsspecfx.header[0]['ns']
+        dt = dobsspecfx.header[0]['dt']/1000000.
+        frqv = np.fft.rfftfreq(ns, dt)
+
+        # Update the SU header
+        dobsspecfx.header[:]['ns'] = len(frqv)
+        dobsspecfx.header[:]['d1'] = frqv[1]-frqv[0]
+        dobsspecfx.header[:]['dt'] = 0
+        dobsspecfx.header[:]['trid'] = 118 # Amplitude of complex trace from 0 to Nyquist
+
+        return dobsspecfx
+
+    def specfk(self):
+        """
+        FK spectrum of traces using the numpy.fft functions.
+        """
+
+        # Create a copy of the input SU data
+        dobsspecfk = copy.deepcopy(self)
+
+        # Get dx from header, if not set dx=1.0
+        d2 = dobsspecfk.header[0]['d2']
+        if d2 == 0:
+            d2 = 1.0
+        ntrac = len(dobsspecfk.header)
+
+        # Amplitude of the real Fourier transform
+        dobsspecfk.trace = np.fft.rfft(self.trace, axis=1)
+        dobsspecfk.trace = np.fft.fft(dobsspecfk.trace, axis=0)
+        dobsspecfk.trace = np.flip(np.fft.fftshift(dobsspecfk.trace, axes=0), axis=0)
+        dobsspecfk.trace = np.absolute(dobsspecfk.trace)
+
+        # Get the frequency and K vectors
+        ns = dobsspecfk.header[0]['ns']
+        dt = dobsspecfk.header[0]['dt']/1000000.
+        frqv = np.fft.rfftfreq(ns, dt)
+        wavv = np.fft.fftfreq(ntrac, d2)
+        # Centering
+        wavv = np.fft.fftshift(wavv)
+
+        # Update the SU header
+        dobsspecfk.header[:]['ns'] = len(frqv)
+        dobsspecfk.header[:]['d1'] = frqv[1]-frqv[0]
+        dobsspecfk.header[:]['d2'] = np.abs(wavv[1]-wavv[0])
+        dobsspecfk.header[:]['dt'] = 0
+        dobsspecfk.header[:]['f1'] = frqv[1]-frqv[0]
+        dobsspecfk.header[:]['f2'] = wavv[0]
+        dobsspecfk.header[:]['trid'] = 122 # Amplitude of complex trace from 0 to Nyquist
+
+        return dobsspecfk